+++ app/listitem/mocks/listitemservice.go
@@ -129,7 +129,7 @@ func (_c *MockListItemService_DeleteListItem_Call) RunAndReturn(run func(context
 }
 
 // FetchAllItemsFromList provides a mock function with given fields: ctx, listID, limit
-func (_m *MockListItemService) FetchAllItemsFromList(ctx context.Context, listID int64, limit int64) ([]listitem.ListItem, error) {
+func (_m *MockListItemService) FetchAllItemsFromList(ctx context.Context, listID int64, limit int32) ([]listitem.ListItem, error) {
 	ret := _m.Called(ctx, listID, limit)
 
 	if len(ret) == 0 {
@@ -138,10 +138,10 @@ func (_m *MockListItemService) FetchAllItemsFromList(ctx context.Context, listID
 
 	var r0 []listitem.ListItem
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, int64, int64) ([]listitem.ListItem, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, int32) ([]listitem.ListItem, error)); ok {
 		return rf(ctx, listID, limit)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, int64, int64) []listitem.ListItem); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, int32) []listitem.ListItem); ok {
 		r0 = rf(ctx, listID, limit)
 	} else {
 		if ret.Get(0) != nil {
@@ -149,7 +149,7 @@ func (_m *MockListItemService) FetchAllItemsFromList(ctx context.Context, listID
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, int64, int64) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, int64, int32) error); ok {
 		r1 = rf(ctx, listID, limit)
 	} else {
 		r1 = ret.Error(1)
@@ -166,14 +166,14 @@ type MockListItemService_FetchAllItemsFromList_Call struct {
 // FetchAllItemsFromList is a helper method to define mock.On call
 //   - ctx context.Context
 //   - listID int64
-//   - limit int64
+//   - limit int32
 func (_e *MockListItemService_Expecter) FetchAllItemsFromList(ctx interface{}, listID interface{}, limit interface{}) *MockListItemService_FetchAllItemsFromList_Call {
 	return &MockListItemService_FetchAllItemsFromList_Call{Call: _e.mock.On("FetchAllItemsFromList", ctx, listID, limit)}
 }
 
-func (_c *MockListItemService_FetchAllItemsFromList_Call) Run(run func(ctx context.Context, listID int64, limit int64)) *MockListItemService_FetchAllItemsFromList_Call {
+func (_c *MockListItemService_FetchAllItemsFromList_Call) Run(run func(ctx context.Context, listID int64, limit int32)) *MockListItemService_FetchAllItemsFromList_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(int64), args[2].(int64))
+		run(args[0].(context.Context), args[1].(int64), args[2].(int32))
 	})
 	return _c
 }
@@ -183,7 +183,7 @@ func (_c *MockListItemService_FetchAllItemsFromList_Call) Return(_a0 []listitem.
 	return _c
 }
 
-func (_c *MockListItemService_FetchAllItemsFromList_Call) RunAndReturn(run func(context.Context, int64, int64) ([]listitem.ListItem, error)) *MockListItemService_FetchAllItemsFromList_Call {
+func (_c *MockListItemService_FetchAllItemsFromList_Call) RunAndReturn(run func(context.Context, int64, int32) ([]listitem.ListItem, error)) *MockListItemService_FetchAllItemsFromList_Call {
 	_c.Call.Return(run)
 	return _c
 }
