+++ app/mocks/services.go
@@ -84,7 +84,7 @@ func (_c *MockServices_AddItemToList_Call) RunAndReturn(run func(context.Context
 }
 
 // AllLists provides a mock function with given fields: ctx, limit
-func (_m *MockServices) AllLists(ctx context.Context, limit int64) ([]list.List, error) {
+func (_m *MockServices) AllLists(ctx context.Context, limit int32) ([]list.List, error) {
 	ret := _m.Called(ctx, limit)
 
 	if len(ret) == 0 {
@@ -93,10 +93,10 @@ func (_m *MockServices) AllLists(ctx context.Context, limit int64) ([]list.List,
 
 	var r0 []list.List
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, int64) ([]list.List, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int32) ([]list.List, error)); ok {
 		return rf(ctx, limit)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, int64) []list.List); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int32) []list.List); ok {
 		r0 = rf(ctx, limit)
 	} else {
 		if ret.Get(0) != nil {
@@ -104,7 +104,7 @@ func (_m *MockServices) AllLists(ctx context.Context, limit int64) ([]list.List,
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, int64) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, int32) error); ok {
 		r1 = rf(ctx, limit)
 	} else {
 		r1 = ret.Error(1)
@@ -120,14 +120,14 @@ type MockServices_AllLists_Call struct {
 
 // AllLists is a helper method to define mock.On call
 //   - ctx context.Context
-//   - limit int64
+//   - limit int32
 func (_e *MockServices_Expecter) AllLists(ctx interface{}, limit interface{}) *MockServices_AllLists_Call {
 	return &MockServices_AllLists_Call{Call: _e.mock.On("AllLists", ctx, limit)}
 }
 
-func (_c *MockServices_AllLists_Call) Run(run func(ctx context.Context, limit int64)) *MockServices_AllLists_Call {
+func (_c *MockServices_AllLists_Call) Run(run func(ctx context.Context, limit int32)) *MockServices_AllLists_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(int64))
+		run(args[0].(context.Context), args[1].(int32))
 	})
 	return _c
 }
@@ -137,7 +137,7 @@ func (_c *MockServices_AllLists_Call) Return(_a0 []list.List, _a1 error) *MockSe
 	return _c
 }
 
-func (_c *MockServices_AllLists_Call) RunAndReturn(run func(context.Context, int64) ([]list.List, error)) *MockServices_AllLists_Call {
+func (_c *MockServices_AllLists_Call) RunAndReturn(run func(context.Context, int32) ([]list.List, error)) *MockServices_AllLists_Call {
 	_c.Call.Return(run)
 	return _c
 }
@@ -294,7 +294,7 @@ func (_c *MockServices_DeleteListItem_Call) RunAndReturn(run func(context.Contex
 }
 
 // FetchAllItemsFromList provides a mock function with given fields: ctx, listID, limit
-func (_m *MockServices) FetchAllItemsFromList(ctx context.Context, listID int64, limit int64) ([]listitem.ListItem, error) {
+func (_m *MockServices) FetchAllItemsFromList(ctx context.Context, listID int64, limit int32) ([]listitem.ListItem, error) {
 	ret := _m.Called(ctx, listID, limit)
 
 	if len(ret) == 0 {
@@ -303,10 +303,10 @@ func (_m *MockServices) FetchAllItemsFromList(ctx context.Context, listID int64,
 
 	var r0 []listitem.ListItem
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, int64, int64) ([]listitem.ListItem, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, int32) ([]listitem.ListItem, error)); ok {
 		return rf(ctx, listID, limit)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, int64, int64) []listitem.ListItem); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, int32) []listitem.ListItem); ok {
 		r0 = rf(ctx, listID, limit)
 	} else {
 		if ret.Get(0) != nil {
@@ -314,7 +314,7 @@ func (_m *MockServices) FetchAllItemsFromList(ctx context.Context, listID int64,
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, int64, int64) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, int64, int32) error); ok {
 		r1 = rf(ctx, listID, limit)
 	} else {
 		r1 = ret.Error(1)
@@ -331,14 +331,14 @@ type MockServices_FetchAllItemsFromList_Call struct {
 // FetchAllItemsFromList is a helper method to define mock.On call
 //   - ctx context.Context
 //   - listID int64
-//   - limit int64
+//   - limit int32
 func (_e *MockServices_Expecter) FetchAllItemsFromList(ctx interface{}, listID interface{}, limit interface{}) *MockServices_FetchAllItemsFromList_Call {
 	return &MockServices_FetchAllItemsFromList_Call{Call: _e.mock.On("FetchAllItemsFromList", ctx, listID, limit)}
 }
 
-func (_c *MockServices_FetchAllItemsFromList_Call) Run(run func(ctx context.Context, listID int64, limit int64)) *MockServices_FetchAllItemsFromList_Call {
+func (_c *MockServices_FetchAllItemsFromList_Call) Run(run func(ctx context.Context, listID int64, limit int32)) *MockServices_FetchAllItemsFromList_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(int64), args[2].(int64))
+		run(args[0].(context.Context), args[1].(int64), args[2].(int32))
 	})
 	return _c
 }
@@ -348,7 +348,7 @@ func (_c *MockServices_FetchAllItemsFromList_Call) Return(_a0 []listitem.ListIte
 	return _c
 }
 
-func (_c *MockServices_FetchAllItemsFromList_Call) RunAndReturn(run func(context.Context, int64, int64) ([]listitem.ListItem, error)) *MockServices_FetchAllItemsFromList_Call {
+func (_c *MockServices_FetchAllItemsFromList_Call) RunAndReturn(run func(context.Context, int64, int32) ([]listitem.ListItem, error)) *MockServices_FetchAllItemsFromList_Call {
 	_c.Call.Return(run)
 	return _c
 }
